splits aiohttp and fastapi parts
