+++ services/payments/src/simcore_service_payments/services/postgres.py
@@ -1,10 +1,5 @@
-import time
-from datetime import timedelta
-
 from fastapi import FastAPI
-from models_library.healthchecks import IsNonResponsive, IsResponsive, LivenessResult
-from servicelib.fastapi.db_async_engine import close_db_connection, connect_to_db
-from sqlalchemy.exc import SQLAlchemyError
+from servicelib.fastapi.db_asyncpg_engine import close_db_connection, connect_to_db
 from sqlalchemy.ext.asyncio import AsyncEngine
 
 from ..core.settings import ApplicationSettings
@@ -14,18 +9,6 @@ def get_engine(app: FastAPI) -> AsyncEngine:
     assert app.state.engine  # nosec
     engine: AsyncEngine = app.state.engine
     return engine
-
-
-async def check_postgres_liveness(engine: AsyncEngine) -> LivenessResult:
-    try:
-        tic = time.time()
-        # test
-        async with engine.connect():
-            ...
-        elapsed_time = time.time() - tic
-        return IsResponsive(elapsed=timedelta(seconds=elapsed_time))
-    except SQLAlchemyError as err:
-        return IsNonResponsive(reason=f"{err}")
 
 
 def setup_postgres(app: FastAPI):
